Move changes to cmd/options.go
