+++ internal/cmd/controller/gitops/operator.go
@@ -10,7 +10,6 @@ import (
 
 	command "github.com/rancher/fleet/internal/cmd"
 	"github.com/rancher/fleet/internal/cmd/controller/gitops/reconciler"
-	"github.com/rancher/fleet/internal/cmd/controller/options"
 	"github.com/rancher/fleet/internal/metrics"
 	fleet "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	"github.com/rancher/fleet/pkg/git/poll"
@@ -87,7 +86,7 @@ func (g *GitOperator) Run(cmd *cobra.Command, args []string) error {
 
 	namespace := g.Namespace
 
-	leaderOpts, err := options.NewLeaderElectionOptions()
+	leaderOpts, err := command.NewLeaderElectionOptions()
 	if err != nil {
 		return err
 	}
