+++ internal/cmd/options.go
@@ -1,4 +1,4 @@
-package options
+package cmd
 
 import (
 	"fmt"
